refactored a code to seperate functions from ui
