--- lib/auth/signup.dart
@@ -1,96 +1,0 @@
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
-
-class SignUp extends StatefulWidget {
-  const SignUp({Key? key}) : super(key: key);
-
-  @override
-  State<SignUp> createState() => _SignUpState();
-}
-
-class _SignUpState extends State<SignUp> {
-  FirebaseAuth auth = FirebaseAuth.instance;
-
-  Future<void> signUpAction({required String emailAddress, required String password}) async {
-    try {
-      final credential = await FirebaseAuth.instance.createUserWithEmailAndPassword(
-        email: emailAddress,
-        password: password,
-      );
-    } on FirebaseAuthException catch (e) {
-      if (e.code == 'weak-password') {
-        debugPrint('The password provided is too weak.');
-      } else if (e.code == 'email-already-in-use') {
-        debugPrint('The account already exists for that email.');
-      }
-    } catch (e) {
-      debugPrint('error $e');
-    }
-  }
-
-  Future<void> signInAction({required String emailAddress, required String password}) async {
-    try {
-      final credential = await FirebaseAuth.instance.signInWithEmailAndPassword(
-        email: emailAddress,
-        password: password,
-      );
-    } on FirebaseAuthException catch (e) {
-      if (e.code == 'weak-password') {
-        debugPrint('The password provided is too weak.');
-      } else if (e.code == 'email-already-in-use') {
-        debugPrint('The account already exists for that email.');
-      }
-    } catch (e) {
-      debugPrint('error $e');
-    }
-  }
-
-
-
-
-  String email = '';
-  String password = '';
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.blue,
-        elevation: 8,
-        title: const Text('Sign Up'),
-      ),
-      body: Container(
-        padding: const EdgeInsets.symmetric(vertical: 20, horizontal: 50),
-        child: Form(
-          child: Column(
-            children: <Widget>[
-              TextFormField(
-                onChanged: (val) => setState(() {
-                  email = val;
-                }),
-              ),
-              TextFormField(
-                onChanged: (val) => setState(() {
-                  password = val;
-                }),
-              ),
-              ElevatedButton(
-                onPressed: () async {
-                  await signUpAction(emailAddress: email, password: password);
-                },
-                child: const Text('SignUp'),
-              ),
-              ElevatedButton(
-                onPressed: () async {
-                  await signInAction(emailAddress: email, password: password);
-                },
-                child: const Text('SignIn'),
-              )
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
