+++ lib/models/user.dart
@@ -1,0 +1,4 @@
+class UserModel {
+  final String id;
+  UserModel({required this.id});
+}
